# Enable metrics while running perf tests