@kislaykishore 2025-03-19
> Please add the same flag in continues test also. [ref](https://github.com/GoogleCloudPlatform/gcsfuse/blob/master/perfmetrics/scripts/continuous_test/gcp_ubuntu/build.sh#L65) and [experiment tests](https://github.com/GoogleCloudPlatform/gcsfuse/blob/master/perfmetrics/scripts/continuous_test/gcp_ubuntu/periodic_experiments/build.sh#L82) if required.

It's not needed - metrics are already enabled for these cases via cloud-exporter.